change looping behaviour to be fast enough for animations while still polling for new messages on the second

+++ lights-controller.py
@@ -1,6 +1,7 @@
-import beanstalkc
+import time
 from time import sleep
 import re
+import beanstalkc
 from LPD8806 import *
 
 beanstalk = beanstalkc.Connection(host='localhost', port=14711, parse_yaml=False)
@@ -12,17 +13,18 @@ colors = {
     'white' : Color(255, 255, 255),
 }
 
+# setup long-lived stateful LEDStrip instance
 default_led_count = 32
 led = LEDStrip(default_led_count)
-led.all_off()
 
-def all_off():
+
+def issue_all_off():
     led.fillOff()
 
 
 # update:2:5:6:1.0:green:white:red:blue:red
 # update:2:5:6:1.0:green:white_pulse:red:blue_pulse:red <<--ignore for now
-def update(tokens):
+def issue_update(tokens):
     led.setMasterBrightness(tokens[4])
     start_idx = tokens[1]
     segment_width = tokens[3]
@@ -36,41 +38,46 @@ def update(tokens):
 
 
 # start_build:2:32:1.0
-def start_build(start_idx=0, led_count=32, brightness=1.0):
+def issue_start_build(start_idx=0, led_count=32, brightness=1.0):
     led.setMasterBrightness(brightness)
-
     led.anim_larson_rainbow(2, 0.5)
     led.update()
-
-    sleep(0.03)
 
     #larson scanner (i.e. Cylon Eye or K.I.T.T.) but Rainbow
     # def anim_larson_rainbow(self, tail=2, fade=0.75, start=0, end=0):
 
 
+def issue_current_directive(directive):
+    tokens = directive.split(':')
+    if tokens[0] == 'all_off':
+        issue_all_off()
+
+    elif tokens[0] == 'start_build':
+        issue_start_build(tokens[1], tokens[2], tokens[3])
+
+    elif tokens[0] == 'update':
+        issue_update(tokens)
+
+
 def main():
     try:
+        directive = 'all_off'
+        job = None
+        last_second = time.localtime().tm_sec
         while True:
-            print 'polling..'
-            job = beanstalk.reserve(timeout=0)
-            if job is not None:
-                print "job found with content: %s" % job.body
-                directive = job.body
-                job.delete()
-                tokens = directive.split(':')
+            issue_current_directive(directive)
 
-                if tokens[0] == 'all_off':
-                    all_off()
+            now = time.localtime().tm_sec
+            if now != last_second:
+                last_second = now
+                print 'polling..'
+                job = beanstalk.reserve(timeout=0)
+                if job is not None:
+                    print "job found with content: {0}".format(job.body)
+                    directive = job.body
+                    job.delete()
 
-                elif tokens[0] == 'start_build':
-                    start_build(tokens[1], tokens[2], tokens[3])
-
-                elif tokens[0[ == 'update':
-                    update(tokens)
-
-
-
-            sleep(0.50)
+            sleep(0.03) # loop fast enough for animations ---> this could be altered per directive if reqd
 
     except KeyboardInterrupt:
         print '^C received, shutting down controller'
